[Calcite Engine] Only enable fallback for tests that need to fall back
